backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
pillow==10.2.0
face_recognition==1.3.0
imagehash==4.3.1
numpy==1.26.3
python-magic==0.4.27
exifread==3.0.0
pydantic==2.5.3
alembic==1.13.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
minio==7.2.3
